Everything I had for labor4, but read description!
Did a bunch of experimenting on site in the lab, so this is everything I used during the entire process to pass the task. I think one of the testbenches used the component directly in the end and not the originally intended labor file. Since the tasked was changed on the fly some stuff was just quickly added and so multiple files were created.